Schema table support (#142)
* [xs]: schema table support

* [schema][xs]: handle possible schema information.

In CKAN Classic, we are implementing tableschema spec in resources descriptor using `ckanext-scheming` and `ckanext-composite` extensions. Since we're limited to have only one level of subfields, we're using `resource.fields` property to list field descriptors. In this commit, we're restructuring the descriptor so that we have `resource.schema` that follows tableschema spec.

* [tests][s]: test 'fields' property is handled into 'schema'.

Co-authored-by: Anuar Ustayev <anuar.ustayev@gmail.com>