[datapackage][s]: created a function that can be re-used/shared so that generated 'datapackage.json' is consistent.
We generate 'datapackage' object for showcase (dataset) page.
We also generate it for 'datapackage.json' api route (`/owner/dataset/datapackage.json`).
Problem is that these 2 objects aren't consistent, however, it should be.
So this PR includes a shared function so that both routes use it and 'datapackage' json is identical.